adjustment to typings

+++ src/ADALottoGameWinner.cs
@@ -9,7 +9,7 @@ namespace ADALottoModels
         public string Address { get; set; } = string.Empty;
         public long DrawBlockId { get; set; }
         public DateTime DrawDate { get; set; }
-        public double Prize { get; set; }
+        public long Prize { get; set; }
         public Transaction? RewardTx { get; set; }
     }
 }
